pyproject.toml
[project]
name = "medical-lab-database"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "streamlit (>=1.45.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "streamlit-extras (>=0.7.1,<0.8.0)",
    "streamlit-autorefresh (>=1.0.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
